Quản lý yêu cầu #TravelRequest.

+++ OneTravelApi/Services/ITravelRequestService.cs
@@ -12,18 +12,23 @@ using System.Threading.Tasks;
 
 namespace OneTravelApi.Services
 {
-    public interface IPartnerContactService : IBaseService<PartnerContactResource>
+    public interface ITravelRequestService
     {
-        IActionResult GetByPartner(int partnerId);
+        IActionResult GetAll(int? pageSize = 10, int? pageNumber = 1, string q = null);
+        Task<IActionResult> GetAsync(int id, bool related = false);
+        Task<IActionResult> Create(TravelRequestSaveResource resource);
+        Task<IActionResult> Update(int id, TravelRequestSaveResource resource);
+        Task<IActionResult> Delete(int id);
     }
 
-    public class PartnerContactService : IPartnerContactService
+    public class TravelRequestService : ITravelRequestService
     {
-        private readonly IRepository<PartnerContact> _repository;
+        private readonly IRepository<TravelRequest> _repository;
         private readonly IMapper _mapper;
-        private readonly ILogger<PartnerContactService> _logger;
+        private readonly ILogger<TravelRequestService> _logger;
 
-        public PartnerContactService(IRepository<PartnerContact> repository, IMapper mapper, ILogger<PartnerContactService> logger)
+        public TravelRequestService(IRepository<TravelRequest> repository,
+            IMapper mapper, ILogger<TravelRequestService> logger)
         {
             _repository = repository;
             _mapper = mapper;
@@ -33,20 +38,28 @@ namespace OneTravelApi.Services
         public IActionResult GetAll(int? pageSize, int? pageNumber, string q = null)
         {
 
-            var response = new ListModelResponse<PartnerContactResource>
+            var response = new ListModelResponse<TravelRequestResource>
             {
                 PageSize = (int)pageSize,
                 PageNumber = (int)pageNumber
             };
-            var query = _repository.Query().Skip((response.PageNumber - 1) * response.PageSize)
+            var query = _repository.Query()
+                .Include(x => x.Partner)
+                .Include(x => x.PartnerContact)
+                .Include(x => x.CategoryRequest)
+                .Include(x => x.User)
+                .Include(x => x.CategoryRequestStatus)
+                .Skip((response.PageNumber - 1) * response.PageSize)
                 .Take(response.PageSize).ToList();
 
             if (!string.IsNullOrEmpty(q) && query.Any())
             {
                 q = q.ToLower();
-                query = query.Where(x => x.ContactName.ToLower().Contains(q)
-                || x.Email.ToLower().Contains(q)
-                || x.PositionTitle.ToLower().Contains(q)).ToList();
+                query = query.Where(x => x.RequestName.ToLower().Contains(q)
+                || x.GuestEmail.ToLower().Contains(q)
+                || x.GuestMobile.ToLower().Contains(q)
+                || x.RequestInfo.ToLower().Contains(q)
+                || x.RequestName.ToLower().Contains(q)).ToList();
             }
 
             try
@@ -65,22 +78,44 @@ namespace OneTravelApi.Services
             return response.ToHttpResponse();
         }
 
-        public async Task<IActionResult> GetAsync(int id)
+        public async Task<IActionResult> GetAsync(int id, bool related = false)
         {
-            var response = new SingleModelResponse<PartnerContactResource>();
+            var response = new SingleModelResponse<TravelRequestResource>();
 
             try
             {
-                var entity = await _repository.Query().FirstOrDefaultAsync(x => x.Id == id);
-
-                if (entity == null)
+                if (related)
                 {
-                    response.DidError = true;
-                    response.ErrorMessage = ResponseMessageConstants.NotFound;
-                    return response.ToHttpResponse();
-                }
+                    var entity = await _repository.Query()
+                        .Include(x => x.Partner)
+                        .Include(x => x.PartnerContact)
+                        .Include(x => x.CategoryRequest)
+                        .Include(x => x.User)
+                        .Include(x => x.CategoryRequestStatus)
+                        .FirstOrDefaultAsync(x => x.Id == id);
 
-                response.Model = _mapper.Map(entity, response.Model);
+                    if (entity == null)
+                    {
+                        response.DidError = true;
+                        response.ErrorMessage = ResponseMessageConstants.NotFound;
+                        return response.ToHttpResponse();
+                    }
+
+                    response.Model = _mapper.Map(entity, response.Model);
+                }
+                else
+                {
+                    var entity = await _repository.Query()
+                        .FirstOrDefaultAsync(x => x.Id == id);
+
+                    if (entity == null)
+                    {
+                        response.DidError = true;
+                        response.ErrorMessage = ResponseMessageConstants.NotFound;
+                        return response.ToHttpResponse();
+                    }
+                    response.Model = _mapper.Map(entity, response.Model);
+                }
             }
             catch (Exception ex)
             {
@@ -92,9 +127,9 @@ namespace OneTravelApi.Services
             return response.ToHttpResponse();
         }
 
-        public async Task<IActionResult> Create(PartnerContactResource resource)
+        public async Task<IActionResult> Create(TravelRequestSaveResource resource)
         {
-            var response = new SingleModelResponse<PartnerContactResource>();
+            var response = new SingleModelResponse<TravelRequestResource>();
 
             if (resource == null)
             {
@@ -105,12 +140,12 @@ namespace OneTravelApi.Services
 
             try
             {
-                var entity = new PartnerContact();
+                var entity = new TravelRequest();
                 _mapper.Map(resource, entity);
 
                 var entityAdded = await _repository.AddAsync(entity);
 
-                response.Model = _mapper.Map(entityAdded, resource);
+                response.Model = _mapper.Map(entityAdded, response.Model);
                 response.Message = ResponseMessageConstants.Success;
             }
             catch (Exception ex)
@@ -123,9 +158,9 @@ namespace OneTravelApi.Services
             return response.ToHttpResponse();
         }
 
-        public async Task<IActionResult> Update(int id, PartnerContactResource resource)
+        public async Task<IActionResult> Update(int id, TravelRequestSaveResource resource)
         {
-            var response = new SingleModelResponse<PartnerContactResource>();
+            var response = new SingleModelResponse<TravelRequestResource>();
 
             if (resource == null)
             {
@@ -148,7 +183,7 @@ namespace OneTravelApi.Services
 
                 await _repository.UpdateAsync(entity);
 
-                response.Model = _mapper.Map(entity, resource);
+                response.Model = _mapper.Map(entity, response.Model);
                 response.Message = ResponseMessageConstants.Success;
             }
             catch (Exception ex)
@@ -164,34 +199,23 @@ namespace OneTravelApi.Services
 
         public async Task<IActionResult> Delete(int id)
         {
-            var response = new SingleModelResponse<PartnerContact>();
+            var response = new SingleModelResponse<TravelRequestResource>();
 
             try
             {
-                var entity = await _repository.DeleteAsync(id);
+                var model = await _repository.GetAsync(id);
 
-                response.Model = entity;
+                if (model == null)
+                {
+                    response.DidError = true;
+                    response.ErrorMessage = ResponseMessageConstants.NotFound;
+                    return response.ToHttpResponse();
+                }
+
+                await _repository.DeleteAsync(id);
+
+                response.Model = _mapper.Map(model, response.Model);
                 response.Message = ResponseMessageConstants.Delete;
-            }
-            catch (Exception ex)
-            {
-                response.DidError = true;
-                response.ErrorMessage = ex.Message;
-                _logger.LogError(ex.InnerException.ToString());
-            }
-
-            return response.ToHttpResponse();
-        }
-
-        public IActionResult GetByPartner(int partnerId)
-        {
-            var response = new ListModelResponse<PartnerContactResource>();
-
-            try
-            {
-                var entity = _repository.Query().Where(x => x.IdPartner == partnerId).ToList();
-
-                response.Model = _mapper.Map(entity, response.Model);
             }
             catch (Exception ex)
             {
